drm/rockchip: dw-dp: support force output
When config dp show log as force output mode, it will not get
edid and detect the connect status(to avoid aux error because
sink may be not physical connected). So it should set default
color format and link info before show kernel logo, which will
be used when select the output color format.

Signed-off-by: Zhang Yubing <yubing.zhang@rock-chips.com>
Change-Id: I76ec29cebb101b8629e81eee3d82c93f53cc9102