dma-buf: heaps: cache system_heap buffer attachment
Before a device try to access a dmabuf, it will do
 - dma_buf_attach(dmabuf, dev) to get a attachment
 - dma_buf_map_attachment(attachment, dir) to get a sg table
the sgt can translate to iova, the device can access the dmabuf by the
iova range.

The dma operations(dma_buf_attach/dma_buf_map_attachment) spend time on
the iommu map if the device has iommu domain. To cache the iommu map
result can save the time, i.e it benifit to cache the iova for device.

Since the iova can get from sgt, so it needs to cache the sgt.

First part:
This patch only fix for the dmabuf exported from system_heap allcator.
For the system_heap, the sgt is created by dup_sg_table and mapped by
dma_map_sgtable.

The system_heap has it private buffer and attachment structure, named
 - struct system_heap_buffer
 - struct dma_heap_attachment

The dma_heap_attchments are stored by list in the system_heap_buffer.

Second part:
This patch also fix the operations for the dma_buf_ops, the origin
operation flow is:
  dma_buf_attach           -> kzalloc(attach) + ops->attach + listadd
  dma_buf_map_attachment   -> ops->map_dma_buf
  dma_buf_unmap_attachment -> ops->unmap_dma_buf
  dma_buf_detach           -> listdel + ops->detach + kfree

With this patch, the operation flow is:
fresh dmabuf
  dma_buf_attach           -> kzalloc(attach) + ops->attach/kref_init + listadd
  dma_buf_map_attachment   -> ops->map_dma_buf
  dma_buf_unmap_attachment -> ops->unmap_dma_buf
  dma_buf_detach           -> listdel + ops->detach + kfree
used dmabuf
  dma_buf_attach           -> kzalloc(attach) + listadd /* cached */
  dma_buf_map_attachment   -> kref_get                  /* mapped */
  dma_buf_unmap_attachment -> kref_put
  dma_buf_detach           -> null
  dma_buf_release          -> ops->unmap_dma_buf + listdel + ops->detach + kfree

Signed-off-by: Jianqun Xu <jay.xu@rock-chips.com>
Change-Id: Ie87ed17b06515c0d1341275b47f8008501dc3144